# FORKATA/project_GeoCode/proxy/docs/swagger.yaml
basePath: /
definitions:
  controller.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:1313
info:
  contact: {}
  description: This is a sample server celler server.
  title: Swagger Example API
  version: "1.0"
paths:
  /api/address/geocode:
    post:
      consumes:
      - application/json
      description: get address from geocode
      parameters:
      - description: Get address from geocode
        in: query
        name: query
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Address from Geocode
  /api/address/search:
    post:
      consumes:
      - application/json
      description: get address
      parameters:
      - description: Get address
        in: query
        name: query
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Address Search
  /mycustompath/pprof/allocs:
    get:
      description: Профиль выборочных аллокаций памяти (allocs)
      produces:
      - application/octet-stream
      responses:
        "200":
          description: pprof allocs binary
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: Allocation Samples
      tags:
      - diagnostics
  /mycustompath/pprof/block:
    get:
      description: Профиль ожиданий блокировок (block profile)
      produces:
      - application/octet-stream
      responses:
        "200":
          description: pprof block binary
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: Block Profile
      tags:
      - diagnostics
  /mycustompath/pprof/cmdline:
    get:
      description: cmdline
      produces:
      - text/html
      responses:
        "200":
          description: HTML со списком профилей
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Pprof Cmdline
      tags:
      - diagnostics
  /mycustompath/pprof/goroutine:
    get:
      description: Снимок текущих горутин (goroutine profile)
      produces:
      - application/octet-stream
      responses:
        "200":
          description: pprof goroutine binary
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: Goroutine Snapshot
      tags:
      - diagnostics
  /mycustompath/pprof/heap:
    get:
      description: Снимок кучи (heap profile)
      produces:
      - application/octet-stream
      responses:
        "200":
          description: pprof heap binary
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: Heap Profile
      tags:
      - diagnostics
  /mycustompath/pprof/index:
    get:
      description: Все доступные pprof-профили
      produces:
      - text/html
      responses:
        "200":
          description: HTML со списком профилей
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Pprof Index
      tags:
      - diagnostics
  /mycustompath/pprof/mutex:
    get:
      description: Профиль ожидания мьютексов (mutex profile)
      produces:
      - application/octet-stream
      responses:
        "200":
          description: pprof mutex binary
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: Mutex Profile
      tags:
      - diagnostics
  /mycustompath/pprof/profile:
    get:
      description: Снимает CPU-профиль за заданное количество секунд (параметр seconds)
      parameters:
      - default: 30
        description: Продолжительность профилирования в секундах
        in: query
        name: seconds
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: pprof CPU binary
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: CPU-профиль
      tags:
      - diagnostics
  /mycustompath/pprof/symbol:
    get:
      description: Выполняет преобразование адресов в имена символов (symbol lookup)
      parameters:
      - description: Адрес или символ для поиска
        in: query
        name: symbol
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: pprof symbol binary
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: Symbol Lookup
      tags:
      - diagnostics
  /mycustompath/pprof/threadcreate:
    get:
      description: Профиль создания системных потоков (threadcreate profile)
      produces:
      - application/octet-stream
      responses:
        "200":
          description: pprof threadcreate binary
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: Thread Creation Profile
      tags:
      - diagnostics
  /mycustompath/pprof/trace:
    get:
      description: Снимает трассировку работы рантайма за заданное количество секунд
        (параметр seconds)
      parameters:
      - default: 1
        description: Длительность трассировки в секундах
        in: query
        name: seconds
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: pprof trace binary
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: Runtime Trace
      tags:
      - diagnostics
swagger: "2.0"